Hi again.
Soo, adding libnotify support to gnunet-gtk :)

Im guessing its not so cleanly done(depending on how you wish to 
structure it, but ill leave that up to the GNUnet developers ;)), but 
its rather small so you shouldn't have any trouble deciding if it should 
go in SVN or not)

Im unsure about the configure-ac.diff, if it is the "right thing(TM)"
(it works here with and without libnotify)
would be nice to be able to disable it altogether.
(i can't find a good autotool tutorial online)

the small diffs are included in the attached 
file(three-diffs-libnotify.tar.gz).

anyways, im of to bed, good night.

/Nicklas Larsson



git-svn-id: https://gnunet.org/svn/GNUnet@2610 140774ce-b5e7-0310-ab8b-a85725594a96