Hi,
attached you find a patch (against head) that fixes the gnunet-search
manual page and the --help output of gnunet-pseudonym.

BTW I somewhat dislike the current implementation of keyword parsing in
gnunet-search: if there is only one argument, the argument is broken at
spaces, if there is more than one arguments, each argument is taken
(almost) literally as a keyword.

$ gnunet-search "Das Kapital"

searches for ("Das" OR "Kapital")

while

$ gnunet-search "Marx" "Das Kapital" 

searches for ("Marx" OR "Das Kapital")

I'd suggest to not change argument treatment based on the number of
arguments, and never break arguments at spaces.  However, for shell
scripts etc it might be nice to have an option to specify a single
keyword string to be broken.  Why not support that via specific option?

e.g.

$ gnunet-search --search-string="Das Kapital"

to search for ("Das" OR "Kapital")

If you're ok with it, I could submit a patch that implements such a fix
and updates the man-page once again.

cheers,

David


git-svn-id: https://gnunet.org/svn/GNUnet@8453 140774ce-b5e7-0310-ab8b-a85725594a96